delete 2 files and update 4 files

+++ script.js
@@ -4,9 +4,17 @@ const input1=document.getElementById("input1")
 const input2=document.getElementById("input2")
 const table=document.getElementById("table")
 const popular=document.getElementById("popular")
+const spinner=document.getElementById("spinner")
+const main=document.querySelector("main")
 
 const random=async()=>{
-const res=await fetch(`https://api.spoonacular.com/recipes/random?number=12&tags=vegetarian,dessert&apiKey=e17c67078f944c33989fb7b5b4b20f2e`)
+  main.style.display="none"
+  spinner.classList.remove("hidden")
+const res=await fetch(`https://api.spoonacular.com/recipes/random?number=12&tags=vegetarian,dessert&apiKey=4a12b45436ee468ab2e129244d4beaaf`)
+if (res.status===402) {
+  alert("Am sorry too many request for today try again tomorrow")
+}
+
 const data=await res.json()
 data.recipes.forEach(food => {
   const {title,image}=food
@@ -19,7 +27,10 @@ data.recipes.forEach(food => {
 });
 
 }
-random()
+random().then(()=>{
+     main.style.display="block"
+  spinner.classList.add("hidden")
+})
 
 
 
@@ -32,13 +43,20 @@ find.onclick = async()=>{
     
     try {
       localStorage.clear()
-      const res= await fetch(`https://api.spoonacular.com/recipes/findByIngredients?ingredients=${ingredient}&number=20&apiKey=061deef803d54c4e8034727f78f93b10`)
+      const res= await fetch(`https://api.spoonacular.com/recipes/findByIngredients?ingredients=${ingredient}&number=20&apiKey=4a12b45436ee468ab2e129244d4beaaf`)
+      if (res.status===402) {
+  alert("Am sorry too many request for today try again tomorrow")
+}
+
     const data=await res.json()
+     if (data.length<1) {
+      return alert("Please insert the Ingredients the correct Way")
+    }
     for (let i = 0; i < data.length; i++) {
       let {id,missedIngredients,usedIngredients,unusedIngredients,title}=data[i]
      localStorage.setItem(i,JSON.stringify({id,missedIngredients,usedIngredients,unusedIngredients,title}))
     }
-    window.location="./show.html"
+    window.location="./find.html"
     } catch (error) {
       console.log(error)
     }
@@ -48,11 +66,19 @@ find.onclick = async()=>{
 show.onclick = async()=>{
     
   try {
+    main.style.display="none"
+  spinner.classList.remove("hidden")
      if(input2.value===""){
-   return alert("Please Insert The Dish Name")
+   return alert("Please insert the Dish name")
    }
-   const res= await fetch(`https://api.spoonacular.com/recipes/complexSearch?query=${input2.value}&number=200&apiKey=061deef803d54c4e8034727f78f93b10`)
+   const res= await fetch(`https://api.spoonacular.com/recipes/complexSearch?query=${input2.value}&number=200&apiKey=4a12b45436ee468ab2e129244d4beaaf`)
+   if (res.status===402) {
+  alert("Am sorry too many request for today try again tomorrow")
+}
     const data=await res.json()
+    if (data.results.length<1) {
+      return alert("Please Insert The Correct Name")
+    }
     localStorage.clear()
     for (let i = 1; i < data.results.length; i++) {
       let sent=input2.value.toUpperCase()
@@ -60,8 +86,11 @@ show.onclick = async()=>{
         localStorage.setItem(i,data.results[i].title) 
        
     }
-  window.location="./food.html"
+  window.location="./show.html"
   } catch (error) {
      alert("Please type the exact food name")
+  }finally{
+    main.style.display="block"
+  spinner.classList.add("hidden")
   }
 }
